+++ Cpp/source/TCPConnection.h
@@ -1,7 +1,7 @@
 /**
 * 
 * Copyright (C) 2006-2009 Anton Gravestam.
-* Copyright (C) 2018-2019 Lennart Andersson.
+* Copyright (C) 2018-2020 Lennart Andersson.
 *
 * This file is part of OPS (Open Publish Subscribe).
 *
@@ -77,8 +77,8 @@ namespace ops
 			return _connected;
 		}
 		
-		virtual void getLocal(Address_T& address, int& port) = 0;
-		virtual void getRemote(Address_T& address, int& port) = 0;
+		virtual void getLocal(Address_T& address, uint16_t& port) = 0;
+		virtual void getRemote(Address_T& address, uint16_t& port) = 0;
 
 		virtual void start() {};
 		virtual void stop() {};
